requirements.txt
art==6.5
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
pillow==12.1.0
requests==2.32.5
urllib3==2.6.3
